fixup! Reorganization ViewModel. Transferred onProfileClickListener.
